refactor(compiler): split compiler.rs into compiler/ module
Breaks the 3631-line compiler.rs into a `compiler/` module organized by
node category, matching the pattern used for the typechecker split in
fed5e61. Public API (`Compiler`, `CompileError`, `compile`,
`compile_named`, `compile_fn_body`) is unchanged; bytecode output for
the conformance suite is byte-identical. Closes #38.

Layout (largest 899 lines, all under the ~1200-line target):
- error.rs — `CompileError`
- yield_scan.rs, pipe.rs — AST helper fns (yield detection, pipe
  placeholder rewriting)
- state.rs — struct definitions, constructors, scope/finally helpers,
  type-alias/schema lowering, top-level declaration ordering, the
  `compile` / `compile_named` / `compile_fn_body` entry points
- expressions.rs — binary/function/method/property/list/dict/interp
- statements.rs — assignment, if/while/for/return/break/continue/guard
- patterns.rs — `compile_destructuring`, `compile_match_expr`
- closures.rs — fn/tool/closure emission
- error_handling.rs — throw / try* / try-catch / try-expr / retry
- concurrency.rs — parallel / spawn / select
- decls.rs — enum/struct constructs, impl blocks, attributed decls,
  `@acp_tool` registration
- mod.rs — thin `compile_node` dispatcher; most arms delegate to a
  per-category helper method in the files above
- tests.rs — existing unit tests
