+++ crates/harn-vm/src/compiler/error.rs
@@ -1,0 +1,14 @@
+/// Compile error.
+#[derive(Debug)]
+pub struct CompileError {
+    pub message: String,
+    pub line: u32,
+}
+
+impl std::fmt::Display for CompileError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "Compile error at line {}: {}", self.line, self.message)
+    }
+}
+
+impl std::error::Error for CompileError {}
